Reduce file sizes (#248)
Various efforts and cleanup with the goal of getting functions.html < 14M. Final solution was to switch to legacy renderer for functions.malloynb since even with aggressive minification our result ModelDefs add up with as many tables as get rendered in that file.

When we retire the legacy renderer we will need to invent a simple static table renderer just for this page.